Switch Client Count display to MVVM
Switch the display of connected clients in the Mainwindow UI to use MVVM
instead of manually updating in MainWindow.xaml.cs

This reduces the amount of things the MainWindow class is responsible
for which makes it less of a God Object and removes the 100ms refresh
times for this element.

It also removes the need for MainWindow class to act as a middleman
between the XAML defined UI and the ConnectedClientsSingleton which is
the authority on this data.