Switch connection status indicators to Bindings
Switch all the connection status image indicators to use Bindings
instead of being set in code.

As part of thise we make the ClientStateSingleton the sole source of
truth for connection status which means added a few new members.

This lets us get rid of some disparate UI update code, including a UI
callback, which overall simplifies the code and is another step in
de-coupling the UI from the application business logic.