Replace if statements with guard clauses
Replace usage of `if` statements at the beginning of methods that
determine whether the rest of the method is run on not with Guard
Clauses.

This reduces the amount of nesting in methods and improves
readibility since you don't need to skip to the bottom of a potentially
long method to determine if an `if` is a Guard Clause or if there is an
`else` down there somewhere.

Process for switching to a guard clause. Find a lop-sided `if` statement
where one of the branches doesn't exist or is vestigial (i.e. contains
no business logic). Then:

1. Invert the conditional
2. Add a new body that returns or continues (Depending on if we are in a
loop or not). Including Debug statements if they were in the old
vestigial branch
3. Remove the brackets around the old big branch
4. Unindent the old branch code