Extract to AudioInputSingleton Class
Create a new `AudionInputSingleton` that acts as the source of truth for
the audio input devices available on the machine SRS is running on and
extract code to this new class.

All checks for microphone availability have been switched from
`ClientStateSingleton` to this new class. This reduces the number of
responsibilities `ClientStateSingleton` has to deal with and makes it
more network client oriented.

As part of this change we will switch the UI elements that rely on this
information to use bindings that access this new class. This gets rid of
more coupling between the business logic and the UI.

Since the availability of audio input devices is checked once at
application startup there is no chance of chance so we don't need to
implement the `INotifyPropertyChanged` interface as we did  on the other
Singleton classes.