Extract to AudioOutputSingleton Class
Create a new `AudionOutputSingleton` that acts as the source of truth
for the audio output devices available on the machine SRS is running on
and extract code to this new class.

As part of this change we will switch the UI elements that rely on this
information to use bindings that access this new class. This gets rid of
more coupling between the business logic and the UI.

We are also making this class the source of truth as to whether the OS
is WindowsN or not. It could be argued that this is a
`ClientStateSingleton` responsibility but since this is only used for
Audio Output related functions (and only two places) I think this class
is the correct place for that.

Since the availability of audio output devices is checked once at
application startup there is no chance of change so we don't need to
implement the `INotifyPropertyChanged` interface.