Remove UI elements in middle of data flow
Remove the use of UI elements as data-sources in the underling business
logic when connecting etc. Instead get the data from the newly extracted
singletons.

This removes the UI from the UI -> Singleton -> UI -> Business logic flow
and instead changes it to UI -> Singleton -> Business Logic which
improves de-coupling the UI from the logic.