Simplify EAM field enabling / disabling
Simplify the code surrounding the enabling and disabling of the
ExternalAWACSMode username and password fields.

This means moving them to a data-bingin on the ClientStateSingleton and
removing all the code in the mainwindow.xaml.cs that was manually
updating the enabled setting.

Doing so reduces a lot of duplication and removes the RedrawUI callback.